requirements.txt
beautifulsoup4==4.8.2
bson==0.5.8
certifi==2019.11.28
dataclasses==0.6
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.0.0
django-admin-material==0.1.0
django-debug-toolbar==2.2
django-heroku==0.3.1
django-htmlmin==0.11.0
django-import-export==2.0.2
django-mobile==0.7.0
django-mongodb-engine==0.6.0
djangorestframework==3.11.0
djangotoolbox==1.8.0
djongo==1.3.1
djongo-mongo-version-fix==1.2.32
dnspython==1.16.0
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
psycopg2==2.8.4
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
sentry-sdk==0.14.2
six==1.14.0
soupsieve==2.0
sqlparse==0.2.4
tablib==1.1.0
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
Pillow